sweet alert added

+++ src/app/components/item/item.component.ts
@@ -5,6 +5,7 @@ import { MatPaginator } from '@angular/material/paginator';
 import { Router } from '@angular/router';
 import { Item } from 'src/app/models/item';
 import { ItemService } from 'src/app/services/item.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-item',
@@ -17,6 +18,7 @@ export class ItemComponent implements OnInit {
 
   itemDataSource!: MatTableDataSource<Item>;
   itemLength!: number;
+  totalItems: number[] = [];
 
   displayedColumns: string[] = ['itemName', 'description', 'price', 'action'];
   itemColumns: string[] = [
@@ -27,8 +29,10 @@ export class ItemComponent implements OnInit {
   ];
 
   @ViewChild(MatSort) sort!: MatSort;
+  totalPrice!: number;
 
   constructor(private itemService: ItemService) { }
+
 
   fetchItems() {
     this.itemService.getAllItems().subscribe({
@@ -45,24 +49,57 @@ export class ItemComponent implements OnInit {
     })
   }
 
+  getTotalItems() {
+    this.itemService.getAllItems().subscribe(results => {
+      this.totalItems = results.data.map(mapitems => {
+        return mapitems.price
+
+      })
+      this.totalPrice = this.totalItems.reduce((a, b) => {
+        return a += b;
+      }, 0)
+
+
+      console.log(this.totalPrice)
+    })
+  }
+
   removeItem(id: string) {
-    this.itemService.deleteService(id).subscribe({
-      next: (res) => {
-        this.fetchItems()
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+        this.itemService.deleteService(id).subscribe({
 
-      }, error: (err) => {
-        console.log(err);
+          next: (res) => {
+            this.fetchItems()
 
+          }, error: (err) => {
+            console.log(err);
+          }
 
+        })
 
       }
-
     })
 
   }
 
   ngOnInit(): void {
     this.fetchItems();
+    this.getTotalItems()
+
   }
 
 }
